Putting together console client, added output and init command to main class.
The RabbitConsole class is coming together pretty well. I have the
basics of output worked out, and the init command is handled so far.

I have a static method for initialising the database in the Rabbit
class. This is definitely preferred to the RabbitBuider class, which was
basically pointless.

The output method is going to need some work, as right now it isn't
dynamic at all, and output would be rather ugly.

I've also mixed import statements into code randomly to improve startup
speed. I've decided this is stupid at the time of writing this message,
and I'm going to move them back and handle it differently.